code/layout-dm/pyproject.toml
[tool.poetry]
name = "trainer"
version = "0.1.0"
description = ""
authors = ["Dummy User <dummy_user@gmail.com>"]
packages = [
    { include = "trainer", from = "src/trainer", format="sdist" },
]

[tool.poetry.dependencies]
python = "3.10.10"
Pillow = "^9.1.0"
google-cloud-storage = "^2.2.1"
tensorflow = "^2.8.0"
hydra-core = "^1.1.2"
einops = "^0.4.1"
tqdm = "^4.64.0"
pandas = "1.3.5"
seaborn = "^0.11.2"
ipython = "<7.32.0"
ipykernel = "^6.13.0"
setuptools = "59.5.0"
Cython = "^0.29.30"
pycocotools = "^2.0.4"
prdc = "^0.2"
scikit-learn = "1.0.2"
numpy = "1.21.6"
matplotlib = "3.5.2"
fsspec = "2023.1.0"
gcsfs = "2023.1.0"
scipy = "1.11.1"
torch = {version = "^2.1.0+cu121", source = "torch_cu121"}
torchvision = {version = "^0.16.0+cu121", source = "torch_cu121"}
torchaudio = {version = "^2.1.0+cu121", source = "torch_cu121"}
torch-geometric = "^2.4.0"
pytorch-fid = "^0.3.0"
omegaconf = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pysen = {extras = ["lint"], version = "^0.10.1"}

[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py37"

[[tool.pysen.lint.mypy_targets]]
  paths = [".", "src/preprocess/preprocess", "src/trainer/trainer"]